+++ modules/controls/wysiwyg/ReactQuill.ts
@@ -10,7 +10,11 @@ export class ReactQuill extends AbstractControls {
             {
                 dev: "/modules/controls/wysiwyg/react-quill-editor.tsx",
                 production: `${this.routPrefix}/assets/modules/react-quill-editor.es.js`
-            }
+            },
+        viewJsPath: {
+            dev: "/modules/controls/wysiwyg/react-quill-editor.tsx",
+            production: `${this.routPrefix}/assets/modules/react-quill-editor.es.js`
+        }
     }
     readonly config: Config = {};
 
@@ -27,6 +31,14 @@ export class ReactQuill extends AbstractControls {
             return this.path.jsPath.dev;
         } else {
             return this.path.jsPath.production
+        }
+    }
+
+    getViewJsPath(): string {
+        if (process.env.VITE_ENV === 'dev') {
+            return this.path.viewJsPath!.dev;
+        } else {
+            return this.path.viewJsPath!.production;
         }
     }